requirements.txt.txt
streamlit
tensorflow
numpy
matplot.lib
pillow